project/python商品销售数据分析可视化系统（带爬虫）/code/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2
django-allauth==0.50.0
django-crispy-forms==1.13.0
django-formtools==2.3
django-import-export==2.7.1
django-reversion==4.0.2
et-xmlfile==1.1.0
future==0.18.2
httplib2==0.9.2
idna==3.3
kiwisolver==1.3.1
MarkupPy==1.14
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.1.0
urllib3==1.26.9
wincertstore==0.2
xadmin-django==3.0.2
xlrd==2.0.1
xlwt==1.3.0
